inferx/models/sam2/_implementations/requirements/requirements_jetson.txt
pillow
opencv-python
matplotlib
numpy==1.26.4
segment-anything
torch-tensorrt
https://developer.download.nvidia.com/compute/redist/jp/v61/pytorch/torch-2.5.0a0+872d972e41.nv24.08.17622132-cp310-cp310-linux_aarch64.whl; python_version == "3.10"
https://github.com/NVIDIA-AI-IOT/torch2trt/archive/refs/heads/master.zip 